+++ packages/ui5-application-writer/src/data/defaults.ts
@@ -1,5 +1,8 @@
 import { App, Package, UI5, UI5Framework } from '../types';
-import mappings from './version-to-descriptor-mapping.json'; // from https://github.com/SAP/ui5-manifest/blob/master/mapping.json
+import versionToManifestDescMapping from './version-to-descriptor-mapping.json'; // from https://github.com/SAP/ui5-manifest/blob/master/mapping.json
+import { getUI5Libs } from './ui5Libs';
+import semVer from 'semver';
+import { t } from '../i18n';
 
 /**
  * Returns a package instance with default properties.
@@ -37,8 +40,8 @@ export function mergeApp(app: App): App {
     return Object.assign(
         {
             version: '0.0.1',
-            title: `Title of ${app.id}`, //todo: localise
-            description: `Description of ${app.id}`, //todo: localise
+            title: t('text.defaultAppTitle', { id: app.id }),
+            description: t('text.defaultAppDescription', { id: app.id }),
             baseComponent: 'sap/ui/core/UIComponent'
         },
         app
@@ -48,43 +51,85 @@ export function mergeApp(app: App): App {
 export enum UI5_DEFAULT {
     DEFAULT_UI5_VERSION = '',
     DEFAULT_LOCAL_UI5_VERSION = '1.95.0',
-    MIN_UI5_VERSION = '1.60',
+    MIN_UI5_VERSION = '1.60.0',
     MIN_LOCAL_SAPUI5_VERSION = '1.76.0',
     MIN_LOCAL_OPENUI5_VERSION = '1.52.5',
     SAPUI5_CDN = 'https://ui5.sap.com',
-    OPENUI5_CDN = 'https://openui5.hana.ondemand.com'
+    OPENUI5_CDN = 'https://openui5.hana.ondemand.com',
+    TYPES_VERSION_SINCE = '1.76.0',
+    TYPES_VERSION_PREVIOUS = '1.71.18',
+    MANIFEST_VERSION = '1.12.0'
 }
+
+// Required default libs
+export const defaultUI5Libs = ['sap.m', 'sap.ui.core'];
 
 /**
  * Merges version properties with the provided UI5 instance.
+ * Coerces provided UI5 versions to valid semantic versions.
  *
  * @param {UI5} [ui5] - the UI5 instance
  * @returns {UI5} the updated copy of UI5 instance (does not change `ui5`)
  */
 export function mergeUi5(ui5: Partial<UI5>): UI5 {
-    const version = ui5.version ?? UI5_DEFAULT.DEFAULT_UI5_VERSION; // no version indicates the latest available should be used
+    const version = ui5.version ?? UI5_DEFAULT.DEFAULT_UI5_VERSION; // Undefined or null indicates the latest available should be used
     const framework = ui5.framework ?? 'SAPUI5';
     const defaultFrameworkUrl = framework === 'SAPUI5' ? UI5_DEFAULT.SAPUI5_CDN : UI5_DEFAULT.OPENUI5_CDN;
     const merged: Partial<UI5> & Pick<UI5, 'minUI5Version' | 'localVersion' | 'version'> = {
-        minUI5Version: ui5.minUI5Version ?? UI5_DEFAULT.MIN_UI5_VERSION,
+        minUI5Version: getMinUI5Version(version, ui5.minUI5Version),
         localVersion: getLocalVersion({ framework, version, localVersion: ui5.localVersion }),
         version,
         framework,
         frameworkUrl: ui5.frameworkUrl ?? defaultFrameworkUrl
     };
-    const typesVersion = parseFloat(merged.localVersion) >= 1.76 ? merged.localVersion : '1.71.18';
+    // typesVersion must be a valid npm semantic version, we know they cannot be null as already validated
+    const localSemVer = semVer.valid(semVer.coerce(merged.localVersion))!;
+    const typesVersion = semVer.gte(localSemVer, UI5_DEFAULT.TYPES_VERSION_SINCE)
+        ? localSemVer
+        : UI5_DEFAULT.TYPES_VERSION_PREVIOUS;
 
-    merged.descriptorVersion =
-        ui5.descriptorVersion ?? (mappings as Record<string, string>)[merged.minUI5Version] ?? '1.12.0';
+    merged.descriptorVersion = getManifestVersion(merged.minUI5Version, ui5.descriptorVersion);
     merged.typesVersion = ui5.typesVersion ?? typesVersion;
     merged.ui5Theme = ui5.ui5Theme ?? 'sap_fiori_3';
-    merged.ui5Libs = ui5.ui5Libs ?? [];
+    merged.ui5Libs = getUI5Libs(ui5.ui5Libs);
 
     return Object.assign({}, ui5, merged) as UI5;
 }
 
-// if a specific local version is provided, use it, otherwise, sync with version but keep minimum versions in mind
 /**
+ * Gets the miminum UI5 version based on the specified version.
+ *
+ * @param ui5Version - the ui5 version
+ * @param minUI5Version - optional minimum ui5 version
+ * @returns minimum UI5 version for manifest
+ */
+function getMinUI5Version(ui5Version: string, minUI5Version?: string) {
+    return minUI5Version ?? (ui5Version ? ui5Version : UI5_DEFAULT.MIN_UI5_VERSION);
+}
+
+/**
+ * Get the manifest descriptor version from the specified miminum UI5 version.
+ * Snapshots are handled by coercion to proper versions.
+ *
+ * @param minUI5Version - the ui5 version to be used to map to the manifest descriptor version
+ * @param manifestVersion - optional manifest descriptor version to be used if provided
+ * @returns - the manifest descriptor version
+ */
+function getManifestVersion(minUI5Version: string, manifestVersion?: string): string {
+    const minUI5SemVer = semVer.coerce(minUI5Version)!;
+    return (
+        manifestVersion ??
+        (minUI5Version &&
+            (versionToManifestDescMapping as Record<string, string>)[
+                `${semVer.major(minUI5SemVer)}.${semVer.minor(minUI5SemVer)}`
+            ]) ??
+        UI5_DEFAULT.MANIFEST_VERSION
+    );
+}
+
+/**
+ * If a specific local version is provided, use it, otherwise, sync with version but keep minimum versions in mind.
+ *
  * @param inputObj input object
  * @param inputObj.framework UI framework
  * @param inputObj.version UI version
@@ -100,19 +145,21 @@ function getLocalVersion({
     version: string;
     localVersion?: string;
 }): string {
-    let result = '';
+    // If explictly provided use, we deliberately do not coerce
     if (localVersion) {
-        result = localVersion;
-    } else {
-        if (version === UI5_DEFAULT.DEFAULT_UI5_VERSION) {
-            result = UI5_DEFAULT.DEFAULT_LOCAL_UI5_VERSION;
-        } else {
-            result =
-                framework === 'SAPUI5' ? UI5_DEFAULT.MIN_LOCAL_SAPUI5_VERSION : UI5_DEFAULT.MIN_LOCAL_OPENUI5_VERSION; // minimum version available as local libs
-        }
-        if (parseFloat(version) > parseFloat(result)) {
-            result = version;
-        }
+        return localVersion;
+    } else if (version === UI5_DEFAULT.DEFAULT_UI5_VERSION) {
+        return UI5_DEFAULT.DEFAULT_LOCAL_UI5_VERSION;
+    }
+
+    let result: string =
+        framework === 'SAPUI5' ? UI5_DEFAULT.MIN_LOCAL_SAPUI5_VERSION : UI5_DEFAULT.MIN_LOCAL_OPENUI5_VERSION; // minimum version available as local libs
+
+    // If the ui5 `version` is higher than the min framework version 'result' then use that as the local version instead
+    // Update to a valid coerced version string e.g. snapshot-1.80 -> 1.80.0. Cannot be null as previously validated.
+    const versionSemVer = semVer.coerce(version)!;
+    if (semVer.gt(versionSemVer, semVer.coerce(result)!)) {
+        result = semVer.valid(versionSemVer)!;
     }
     return result;
 }