+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 0.10.25
+
+### Patch Changes
+
+-   Updated dependencies [282b6af]
+    -   @sap-ux/ui5-application-writer@0.10.10
+
+## 0.10.24
+
+### Patch Changes
+
+-   Updated dependencies [11909e1]
+    -   @sap-ux/ui5-application-writer@0.10.9
+
 ## 0.10.23
 
 ### Patch Changes